Fix dependencies
Move from depenencies to devDependencies:
- eslint-plugin-react-hooks

Add to devDependencies:
- @types/react

In a React + TypeScript project, the @types/react is necessary for type
checking to work properly. However, adding it to
@ultimaker/react-web-components breaks all tests that use the button.tsx
component because this component allows a type="link" on a button that
does not exist.

This component will have to be refactored into a button.tsx component
that outputs a HTML <button> element and a button-link.tsx component
that outputs a HTML <a> element with a href attribute.