Add new InputText component
<InputText /> is a highly re-usable leaf component that renders a humble
<input class="input-text" type="text" /> element.

It accepts all attributes a normal <input> element can have as props but
will always have the HTML class attribte class="input-text".

The 'type' attribute is restricted to 'text', 'password', 'email', 'url'
and 'number'.

It also accepts 'hasError' as a boolean prop to visually indicate an
error.

Finally, it is possible for a parent to forward a ref to the <input>
element that can be used, for example, for focusing.