Update packages
All packages that could be updated without tests or the build failing
were updated to the last versions.

Updating eslint-plugin-react required a few fixes where code was
violating the linting rules.

Updating to the latest version of Storybook is not trivial. Currently
@dump247/storybook-state is used, but this package does not work with
the next major version of Storybook. This, in addition to some typing
issues, is the biggest hurdle to upgrade to the latest version of
Storybook.